Added all the relationships of the tables

+++ ecommerce-backend/ecommerce-backend/src/main/java/com/team6/ecommercebackend/repositries/OrderDetailsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.team6.ecommercebackend.entities.OrderDetails;
 
 @Repository
-public interface OrderDetailsRepository extends CrudRepository<OrderDetails, Integer>{
+public interface OrderDetailsRepository extends CrudRepository<OrderDetails, Long>{
 
 }
